feat: add token kinds and update compiler pipeline for new numeric types
- Add token kinds for int8-128, uint8-128, float8-128, byte, char
- Update lexer to recognize new type keywords
- Update parser to handle new type tokens
- Update type checker with new numeric type handling
- Update IR types to support all new numeric types
- Update codegen to handle new IR types
- Update integration tests to use int64 instead of deprecated int

+++ tests/integration_test.rs
@@ -14,7 +14,7 @@ fn test_parser_basic() {
     use xin_parser::Parser;
 
     let source = r#"
-        func add(a: int, b: int) int {
+        func add(a: int64, b: int64) int64 {
             return a + b
         }
     "#;
@@ -53,7 +53,7 @@ fn test_ir_builder_basic() {
     use xin_ir::IRBuilder;
 
     let source = r#"
-        func add(a: int, b: int) int {
+        func add(a: int64, b: int64) int64 {
             return a + b
         }
     "#;
